Proof-reading docstrings after refactor.
