+++ tests/lots/test_add_lot.py
@@ -1,5 +1,8 @@
-import pytest, copy
-from chemical_inventory_api_v1 import LotSchema, ValidationErrorCodes, ChemicalSchema
+import copy
+
+from app.utils.validation_error_codes import ValidationErrorCodes
+from app.models.chemicals import ChemicalSchema
+from app.models.lots import LotSchema
 
 # Variable names for Water, In-House are verbose to avoid potential future conflict
 # with purchased water reagents used in other tests.
