+++ tests/lots/test_get_all_lots.py
@@ -1,5 +1,3 @@
-import copy
-from chemical_inventory_api_v1 import LotSchema
 from tests.lots.conftest import post_all_lots
 
 # Variable names for Water, In-House are verbose to avoid potential future conflict
